@@ -3,6 +3,6 @@
 }
 
 <div class="text-center">
-    <h1 class="display-4">Welcome</h1>
-    <p>Learn about <a href="https://docs.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
+    <h1 class="display-4">Second Hand Web</h1>
+    <p>Learn about <a href="https://docs.microsoft.com/aspnet/core">bla bla bla</a>.</p>
 </div>
